+++ examples/lissajous.py
@@ -21,23 +21,39 @@ plot = mp.vstack(
     mp.text("BROWNIAN MOTION"),
     mp.text("first two dimensions"),
     mp.scatter(
-        traj[:, :2],
+        xs=traj[:, 0],
+        ys=traj[:, 1],
         color=(.5,.5,.5),
         height=20,
         width=75,
     ),
     mp.text("principal components 1, 2, 3 over time"),
-    mp.scatter(np.c_[time, proj[:,0]], color="red", width=75)
-    @ mp.scatter(np.c_[time, proj[:,1]], color="green", width=75)
-    @ mp.scatter(np.c_[time, proj[:,2]], color="blue", width=75),
+    mp.scatter(xs=time, ys=proj[:,0], color="red", width=75)
+    @ mp.scatter(xs=time, ys=proj[:,1], color="green", width=75)
+    @ mp.scatter(xs=time, ys=proj[:,2], color="blue", width=75),
     mp.text("principal components 1, 2, 3 paired"),
     mp.hstack(
         mp.text("PC1 v PC2")
-            / mp.scatter(proj[:, (0,1)], color="yellow", width=25)
-        + mp.text("PC1 v PC3")
-            / mp.scatter(proj[:, (0,2)], color="magenta", width=25)
-        + mp.text("PC2 v PC3")
-            / mp.scatter(proj[:, (1,2)], color="cyan", width=25)
+        / mp.scatter(
+            xs=proj[:, 0],
+            ys=proj[:, 1],
+            color="yellow",
+            width=25,
+        ),
+        mp.text("PC1 v PC3")
+        / mp.scatter(
+            xs=proj[:, 0],
+            ys=proj[:, 2],
+            color="magenta",
+            width=25,
+        ),
+        mp.text("PC2 v PC3")
+        / mp.scatter(
+            xs=proj[:, 1],
+            ys=proj[:, 2],
+            color="cyan",
+            width=25,
+        )
     ),
 )